# static/openapi.yaml
openapi: 3.0.3
info:
  title: AndroBlight API
  description: |
    **AndroBlight** — Android Malware Detection API (CNN-BiLSTM v3)

    Analyse APK files and Play Store packages for malware using a CNN-BiLSTM
    deep-learning model, multi-engine AV enrichment (VirusTotal, Hybrid Analysis,
    MetaDefender), permission-risk scoring, and custom threat rules.

    ## Authentication
    Most endpoints require a **Bearer JWT** token obtained from `POST /auth/login`.
    Include it in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
  version: "3.1.0"
  contact:
    name: AndroBlight Group-47
servers:
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: user@example.com
        display_name:
          type: string
          example: Alice
        created_at:
          type: string
          format: date-time

    # ── Scan result ───────────────────────────────────────────────────────────
...
paths:
...
  /auth/2fa/status:
    get:
      tags: [Auth]
      summary: Get 2FA enabled status for the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 2FA status
          content:
            application/json:
              schema:
                type: object
                properties:
                  two_factor_enabled:
                    type: boolean
        "401":
          description: Not authenticated
  /auth/2fa/setup:
    post:
      tags: [Auth]
      summary: Generate a new TOTP secret and QR provisioning URI
      description: >
        Returns the base32 secret and an `otpauth://` URI suitable for QR display.
        2FA is **not** activated until `/auth/2fa/confirm` is called with a valid OTP.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: TOTP secret generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                    example: "JBSWY3DPEHPK3PXP"
                  otpauth_uri:
                    type: string
                    example: "otpauth://totp/AndroBlight:user@example.com?secret=JBSWY3DP&issuer=AndroBlight"
        "401":
          description: Not authenticated
  /auth/2fa/confirm:
    post:
      tags: [Auth]
      summary: Confirm and activate 2FA with first OTP
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp]
              properties:
                otp:
                  type: string
                  minLength: 6
                  maxLength: 6
                  example: "123456"
      responses:
        "200":
          description: 2FA activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Invalid OTP or 2FA not set up
        "401":
          description: Not authenticated
  /auth/2fa/verify:
    post:
      tags: [Auth]
      summary: Complete login by verifying OTP (2FA step 2)
      description: >
        Public endpoint. Requires `pre_auth_token` returned from `POST /auth/login`
        when 2FA is enabled.  Returns full `access_token` + `refresh_token` on success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pre_auth_token, otp]
              properties:
                pre_auth_token:
                  type: string
                otp:
                  type: string
                  minLength: 6
                  maxLength: 6
                  example: "654321"
      responses:
        "200":
          description: Login complete — full JWT tokens returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Invalid or expired pre_auth_token or wrong OTP
  /auth/2fa/disable:
    post:
      tags: [Auth]
      summary: Disable 2FA (requires current password and OTP)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, otp]
              properties:
                password:
                  type: string
                otp:
                  type: string
                  minLength: 6
                  maxLength: 6
      responses:
        "200":
          description: 2FA disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Invalid OTP or wrong password
        "401":
          description: Not authenticated

  # ── Scan ─────────────────────────────────────────────────────────────────────
...